refactor: Enhance routes validation handling
